move vml work file location for easier browsing.
