+++ tasks/pmm_vml/__init__.py
@@ -1,0 +1,4 @@
+"""
+PMM VML (Pure Market Making with Variable Machine Learning) module.
+This module contains tasks related to the PMM VML strategy.
+"""
